Fix dependencies when $manage_usr_grp=false
In various places the dependencies are written making the assumption
$manage_usr_grp is always true. When set to false, the User[$user] resource
is not in the catalog and may not be referenced.

This commit separates out the dependencies on User[$user] using -> and ~>
operators, so that they're only included in the catalog if $manage_usr_grp
is enabled.

Tested on CentOS6.7 and bitbucket 4.6.