Issue 44: EL init process pattern not matched in 5
This commit updates the EL init script template to use a different
process pattern for BitBucket version 5. Prior to this, the catalina
base was searched for, which doesn't appear to be present in the 5.x
command arguments.

It also updates the pattern for < 5 to grep the java namespace rather
than the catalina base.